LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/container-units-svglength-expected.txt

CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="calc(10cqmin + 10cqmax)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="calc(10cqw + 3px)"

PASS unitType with container-relative units
PASS cqw,cqh can be resolved
PASS cqi,cqb can be resolved
PASS cqmin,cqmax can be resolved
FAIL calc() with container-relative units can be resolved assert_equals: expected 35 but got 0
FAIL Can modify value with container-relative units assert_equals: expected "80" but got "40cqw"
PASS CSS Container Queries Test: container-relative units in SVGLength
